Game over screen, win condition, refactoring and more
Added a game over screen, and a win condition if the player reaches the end of the path. Refactored code to make it more centralized and easier to read. Trail effect on player ship. Refactored laser pool system so it is split between player and enemies. Made build for third demo fest.